requirements.txt
six
chardet
requests
selenium
beautifulsoup4
attrs
fake-useragent
diskcache<3.0.0
sqlalchemy_mate
pymongo_mate
mongoengine_mate
mongomock_mate
scrapy